+++ tests/test_layers.py
@@ -36,7 +36,7 @@ def test_albedo():
 def test_alos_dsm():
     data = AlosDSM().get_data(BBOX)
     assert np.size(data) > 0
-    assert_raster_stats(data, 1, 0, 66, 1122, 0)
+    assert_raster_stats(data, 1, 0, 68, 1122, 0)
     assert get_projection_type(data.crs) == ProjectionType.UTM
     utm_bbox_data = AlosDSM().get_data(BBOX_AS_UTM)
     assert get_rounded_gdf_geometry(data, 1).equals(get_rounded_gdf_geometry(utm_bbox_data, 1))
@@ -117,7 +117,7 @@ def test_high_land_surface_temperature():
 def test_high_slope():
     data = HighSlope().get_data(BBOX)
     assert np.size(data) > 0
-    assert_raster_stats(data, 1, 10.07, 22.01, 143, 979)
+    assert_raster_stats(data, 1, 10.0, 25.2, 146, 976)
     assert get_projection_type(data.crs) == ProjectionType.UTM
     utm_bbox_data = HighSlope().get_data(BBOX_AS_UTM)
     assert get_rounded_gdf_geometry(data, 1).equals(get_rounded_gdf_geometry(utm_bbox_data, 1))